Updated code to address minor comments - updated comment / refactored helper to env.ts
