+++ libs/typescript/core/src/exporters/wal/index.ts
@@ -1,10 +1,5 @@
 /**
  * Public entry point for the WAL exporter package.
- *
- * Only `MlflowWalSpanExporter` is re-exported here; the daemon,
- * supervisor, storage and clients modules are implementation details
- * that are either bundled into the standalone daemon binary or
- * imported directly by tests, not part of the @mlflow/core public API.
  */
 
 export { MlflowWalSpanExporter } from './exporter';
